mv twitch command
