+++ .vscode/settings.json
@@ -1,9 +1,7 @@
 {
-    "editor.formatOnSave": false,
-    "eslint.validate": [
-        "typescript"
-    ],
-    "editor.codeActionsOnSave": {
-        "source.fixAll": true
-    }
+  "editor.formatOnSave": false,
+  "eslint.validate": ["typescript"],
+  "editor.codeActionsOnSave": {
+    "source.fixAll": true
+  }
 }
