+++ src/dota/events.ts
@@ -5,7 +5,7 @@ import { prisma } from '../db/prisma.js'
 import RedisClient from '../db/redis.js'
 import { DBSettings, getValueOrDefault } from '../db/settings.js'
 import Mongo from '../steam/mongo.js'
-import { chatClient } from '../twitch/commands/index.js'
+import { chatClient } from '../twitch/index.js'
 import { closeTwitchBet } from '../twitch/lib/closeTwitchBet.js'
 import { disabledBets, openTwitchBet } from '../twitch/lib/openTwitchBet.js'
 import { DotaEvent, DotaEventTypes, MapData, Packet, SocketClient } from '../types.js'
