+++ src/twitch/commands/apm.ts
@@ -1,8 +1,7 @@
 import { DBSettings, getValueOrDefault } from '../../db/settings.js'
 import { isPlayingMatch } from '../../dota/lib/isPlayingMatch.js'
+import { chatClient } from '../index.js'
 import commandHandler, { MessageType } from './CommandHandler.js'
-
-import { chatClient } from './index.js'
 
 commandHandler.registerCommand('apm', {
   aliases: [],
