+++ src/twitch/commands/pleb.ts
@@ -1,7 +1,6 @@
 import { DBSettings, getValueOrDefault } from '../../db/settings.js'
+import { chatClient } from '../index.js'
 import commandHandler, { MessageType } from './CommandHandler.js'
-
-import { chatClient } from './index.js'
 
 export const plebMode = new Set()
