+++ src/twitch/commands/refresh.ts
@@ -1,7 +1,6 @@
 import { server } from '../../dota/index.js'
+import { chatClient } from '../index.js'
 import commandHandler, { MessageType } from './CommandHandler.js'
-
-import { chatClient } from './index.js'
 
 commandHandler.registerCommand('refresh', {
   aliases: [],
