+++ tsconfig.json
@@ -1,10 +1,6 @@
 {
-  "include": [
-    "src/**/*"
-  ],
-  "exclude": [
-    "node_modules"
-  ],
+  "include": ["src/**/*"],
+  "exclude": ["node_modules"],
   "compilerOptions": {
     "target": "ESNext",
     "module": "ESNext",
@@ -13,9 +9,7 @@
     "outDir": "./dist",
     "sourceMap": false,
     "strictNullChecks": true,
-    "types": [
-      "node"
-    ],
+    "types": ["node"],
     "allowJs": true,
     "checkJs": true,
     "resolveJsonModule": false,
@@ -28,9 +22,6 @@
   "ts-node": {
     "esm": true,
     "pretty": true,
-    "require": [
-      "dotenv/config",
-      "tsconfig-paths/register"
-    ]
+    "require": ["dotenv/config", "tsconfig-paths/register"]
   }
 }
