+++ tsconfig.production.json
@@ -1,8 +1,8 @@
 {
-    "extends": "./tsconfig",
-    "compilerOptions": {
-      "sourceMap": true,
-      "inlineSources": true,
-      "sourceRoot": "/"
-    }
+  "extends": "./tsconfig",
+  "compilerOptions": {
+    "sourceMap": true,
+    "inlineSources": true,
+    "sourceRoot": "/"
   }
+}
